Move MavenArtifact class to base package since it is part of the public API
