+++ src/main/java/io/github/ascopes/protobufmavenplugin/dependency/BinaryPluginResolver.java
@@ -16,6 +16,7 @@
 
 package io.github.ascopes.protobufmavenplugin.dependency;
 
+import io.github.ascopes.protobufmavenplugin.MavenArtifact;
 import io.github.ascopes.protobufmavenplugin.platform.Digests;
 import io.github.ascopes.protobufmavenplugin.platform.FileUtils;
 import java.io.IOException;
