+++ src/main/java/io/github/ascopes/protobufmavenplugin/dependency/MavenDependencyPathResolver.java
@@ -16,6 +16,7 @@
 
 package io.github.ascopes.protobufmavenplugin.dependency;
 
+import io.github.ascopes.protobufmavenplugin.MavenArtifact;
 import java.io.File;
 import java.nio.file.Path;
 import java.util.ArrayList;
